+++ src/main/kotlin/org/gmd/slack/model/SlackAttachment.kt
@@ -1,3 +1,3 @@
-package org.gmd.slack
+package org.gmd.slack.model
 
 class SlackAttachment(val text: String)
