+++ src/main/kotlin/org/gmd/slack/model/SlackOpenDialog.kt
@@ -1,6 +1,7 @@
-package org.gmd.slack
+package org.gmd.slack.model
 
 import com.fasterxml.jackson.annotation.JsonProperty
+import org.gmd.slack.model.SlackDialog
 import org.gmd.util.JsonUtils.Companion.JSON
 
 class SlackOpenDialog() {
