+++ src/main/kotlin/org/gmd/slack/service/SlackService.kt
@@ -1,6 +1,6 @@
-package org.gmd.service
+package org.gmd.slack.service
 
-import org.gmd.slack.SlackTeamAuth
+import org.gmd.slack.model.SlackTeamAuth
 
 interface SlackService {
